Move Extension data class inside VulnerableExtensionThreatModel
