+++ fluxc/src/main/java/org/wordpress/android/fluxc/model/scan/threat/ThreatMapper.kt
@@ -2,8 +2,6 @@ package org.wordpress.android.fluxc.model.scan.threat
 
 import org.wordpress.android.fluxc.model.scan.threat.ThreatModel.CoreFileModificationThreatModel
 import org.wordpress.android.fluxc.model.scan.threat.ThreatModel.DatabaseThreatModel
-import org.wordpress.android.fluxc.model.scan.threat.ThreatModel.Extension
-import org.wordpress.android.fluxc.model.scan.threat.ThreatModel.Extension.ExtensionType
 import org.wordpress.android.fluxc.model.scan.threat.ThreatModel.FileThreatModel
 import org.wordpress.android.fluxc.model.scan.threat.ThreatModel.Fixable
 import org.wordpress.android.fluxc.model.scan.threat.ThreatModel.Fixable.FixType
@@ -11,6 +9,8 @@ import org.wordpress.android.fluxc.model.scan.threat.ThreatModel.Fixable.FixType
 import org.wordpress.android.fluxc.model.scan.threat.ThreatModel.GenericThreatModel
 import org.wordpress.android.fluxc.model.scan.threat.ThreatModel.ThreatStatus
 import org.wordpress.android.fluxc.model.scan.threat.ThreatModel.VulnerableExtensionThreatModel
+import org.wordpress.android.fluxc.model.scan.threat.ThreatModel.VulnerableExtensionThreatModel.Extension
+import org.wordpress.android.fluxc.model.scan.threat.ThreatModel.VulnerableExtensionThreatModel.Extension.ExtensionType
 import org.wordpress.android.fluxc.network.rest.wpcom.scan.threat.Threat
 import java.util.Date
 import javax.inject.Inject
